+++ src/main/java/googlymod/helpers/GooglyEyeConfig.java
@@ -1,10 +1,17 @@
 package googlymod.helpers;
 
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.URL;
+import java.net.URLClassLoader;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.TreeMap;
 
 import com.badlogic.gdx.Gdx;
+import com.evacipated.cardcrawl.modthespire.Loader;
+import com.evacipated.cardcrawl.modthespire.ModInfo;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.megacrit.cardcrawl.helpers.AsyncSaver;
@@ -29,78 +36,113 @@ public class GooglyEyeConfig {
     public HashMap<String,ArrayList<CardEye>> charSelect = new HashMap<>();
     public HashMap<String,ArrayList<CreatureEye>> creatures = new HashMap<>();
 
-    private static final String DATA_FILE = "googlymod/eye-locations.json";
-    private static GooglyEyeConfig theConfig = null;
+    // Global configuration
+    private static GooglyEyeConfig theConfig;
+    // Edited configuration
+    // Note: use TreeMap to get sorted json files
+    private static class EditedConfig {
+        public TreeMap<String,ArrayList<CardEye>> cards = new TreeMap<>();
+        public TreeMap<String,ArrayList<CardEye>> relics = new TreeMap<>();
+        public TreeMap<String,ArrayList<CardEye>> events = new TreeMap<>();
+        public TreeMap<String,ArrayList<CardEye>> charSelect = new TreeMap<>();
+        //public TreeMap<String,ArrayList<CreatureEye>> creatures = new TreeMap<>();
 
-    private static boolean lotsOfEyes = false;
-    private static ArrayList<CardEye> noEyes;
+        private void save() {
+            AsyncSaver.save("googly-eye-locations.json", gson.toJson(this));
+        }
+    }
+    static EditedConfig editedConfig = new EditedConfig();
+
+    private static ArrayList<CardEye> noEyes = new ArrayList<>();
+
+    private void merge(GooglyEyeConfig config) {
+        cards.putAll(config.cards);
+        relics.putAll(config.relics);
+        events.putAll(config.events);
+        charSelect.putAll(config.charSelect);
+        creatures.putAll(config.creatures);
+    }
+
+    static Gson gson;
     static {
-        noEyes = new ArrayList<>();
-        if (lotsOfEyes) {
-            noEyes.add(new CardEye(250-30, 190, 25));
-            noEyes.add(new CardEye(250+30, 190, 25));
-            noEyes.add(new CardEye(250, 90, 50));
+        GsonBuilder gsonBuilder = new GsonBuilder();
+        gson = gsonBuilder.setPrettyPrinting().create();
+    }
+
+    private static GooglyEyeConfig load(String path) {
+        String contents = Gdx.files.internal(path).readString(String.valueOf(StandardCharsets.UTF_8));
+        return gson.fromJson(contents, GooglyEyeConfig.class);
+    }
+
+    private static GooglyEyeConfig loadFromModJar(URL mod_jar) {
+        URLClassLoader loader = null;
+        try {
+            loader = new URLClassLoader(new URL[] {mod_jar}, null);
+            InputStream in = loader.getResourceAsStream("googly-eye-locations.json");
+            if (in == null) return null;
+            GooglyEyeConfig config = gson.fromJson(new InputStreamReader(in, StandardCharsets.UTF_8), GooglyEyeConfig.class);
+            in.close();
+            return config;
+        } catch (Exception e) {
+            System.out.println(mod_jar);
+            e.printStackTrace();
+            return null;
+        } finally {
+            if (loader != null) {
+                try {
+                    loader.close();
+                } catch(Exception e) {
+                    System.out.println(mod_jar);
+                    e.printStackTrace();
+                }
+            }
         }
     }
 
-    private static void load() {
-        GsonBuilder gsonBuilder = new GsonBuilder();
-        Gson gson = gsonBuilder.create();
-        String contents = Gdx.files.internal(DATA_FILE).readString(String.valueOf(StandardCharsets.UTF_8));
-        theConfig = gson.fromJson(contents, GooglyEyeConfig.class);
-    }
-
-    private static void save() {
-        GsonBuilder gsonBuilder = new GsonBuilder();
-        Gson gson = gsonBuilder.setPrettyPrinting().create();
-        AsyncSaver.save("eye-locations.json", gson.toJson(theConfig));
+    static {
+        theConfig = load("googlymod/eye-locations.json");
+        //theConfig.merge(load("googlymod/eye-locations-conspire.json"));
+        //theConfig.merge(load("googlymod/eye-locations-hubris.json"));
+        // Load googly eye locations from other mods
+        for (ModInfo modinfo : Loader.MODINFOS) {
+            GooglyEyeConfig config = loadFromModJar(modinfo.jarURL);
+            if (config != null) theConfig.merge(config);
+        }
     }
 
     public static ArrayList<CardEye> getCardEyes(String cardId) {
-        if (theConfig == null) load();
-        ArrayList<CardEye> result = theConfig.cards.get(cardId);
-        if (result == null) return noEyes;
-        return result;
+        return theConfig.cards.getOrDefault(cardId, noEyes);
     }
     public static void setCardEyes(String cardId, ArrayList<CardEye> eyes) {
-        if (theConfig == null) load();
         theConfig.cards.put(cardId, eyes);
-        save();
+        editedConfig.cards.put(cardId, eyes);
+        editedConfig.save();
     }
 
     public static ArrayList<CardEye> getRelicEyes(String relicId) {
-        if (theConfig == null) load();
-        ArrayList<CardEye> result = theConfig.relics.get(relicId);
-        if (result == null) return noEyes;
-        return result;
+        return theConfig.relics.getOrDefault(relicId, noEyes);
     }
     public static void setRelicEyes(String relicId, ArrayList<CardEye> eyes) {
-        if (theConfig == null) load();
         theConfig.relics.put(relicId, eyes);
-        save();
+        editedConfig.relics.put(relicId, eyes);
+        editedConfig.save();
     }
 
     public static ArrayList<CardEye> getCharSelectEyes(String playerClass) {
-        if (theConfig == null) load();
-        ArrayList<CardEye> result = theConfig.charSelect.get(playerClass);
-        if (result == null) return noEyes;
-        return result;
+        return theConfig.charSelect.getOrDefault(playerClass, noEyes);
     }
     public static void setCharSelectEyes(String playerClass, ArrayList<CardEye> eyes) {
-        if (theConfig == null) load();
         theConfig.charSelect.put(playerClass, eyes);
-        save();
+        editedConfig.charSelect.put(playerClass, eyes);
+        editedConfig.save();
     }
 
     public static ArrayList<CardEye> getEventEyes(String eventImage) {
-        if (theConfig == null) load();
-        ArrayList<CardEye> result = theConfig.events.get(eventImage);
-        if (result == null) return noEyes;
-        return result;
+        return theConfig.events.getOrDefault(eventImage, noEyes);
     }
     public static void setEventEyes(String eventImage, ArrayList<CardEye> eyes) {
-        if (theConfig == null) load();
         theConfig.events.put(eventImage, eyes);
-        save();
+        editedConfig.events.put(eventImage, eyes);
+        editedConfig.save();
     }
 }